pyproject.toml
[project]
name = "ai-app"
version = "0.1.0"
description = ""
authors = [
    {name = "yuta-2001",email = "yu0210ta.sug@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.44.1,<2.0.0)",
    "openai (>=1.72.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
